+++ application/views/main-frame.php
@@ -13,8 +13,7 @@
       js = d.createElement(s); js.id = id;
       js.src = "//connect.facebook.net/en_US/sdk.js#xfbml=1&version=v2.3";
       fjs.parentNode.insertBefore(js, fjs);
-    }(document, 'script', 'facebook-jssdk'));
-    </script>
+    }(document, 'script', 'facebook-jssdk'));</script>
     <!--load page-->
     <?php $this->load->view('includes/main-header'); ?>
     <div id="content" class="pure-g-r" data-page="<?php echo $page; ?>">